refactor: 카드 캐시 생성 클래스 분리
