+++ src/main/java/blackjack/domain/Card.java
@@ -1,18 +1,8 @@
 package blackjack.domain;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Objects;
 
 public class Card {
-
-    private static final List<Card> CACHE = new ArrayList<>();
-
-    static {
-        for (CardSuit suit : CardSuit.values()) {
-            generateCard(suit);
-        }
-    }
 
     private final CardSuit cardSuit;
     private final CardNumber cardNumber;
@@ -20,16 +10,6 @@ public class Card {
     public Card(CardSuit cardSuit, CardNumber cardNumber) {
         this.cardSuit = cardSuit;
         this.cardNumber = cardNumber;
-    }
-
-    private static void generateCard(CardSuit suit) {
-        for (CardNumber number : CardNumber.values()) {
-            CACHE.add(new Card(suit, number));
-        }
-    }
-
-    public static List<Card> values() {
-        return List.copyOf(CACHE);
     }
 
     public CardSuit getSuit() {
